refactor: move some rust code into its own files and use a tryfrom for context materialization
