+++ pure-wasm/src/serialisation.rs
@@ -1,0 +1,176 @@
+use std::cell::RefCell;
+
+use flatbuffers::{FlatBufferBuilder, Follow, WIPOffset};
+use unleash_types::client_metrics::MetricBucket;
+use unleash_yggdrasil::{ExtendedVariantDef, ToggleDefinition};
+
+use crate::messaging::messaging::{
+    FeatureDefBuilder, FeatureDefs, FeatureDefsBuilder, MetricsBucket, MetricsBucketBuilder, Response, ResponseBuilder, ToggleEntryBuilder, ToggleStatsBuilder, Variant, VariantBuilder, VariantEntryBuilder, VariantPayloadBuilder
+};
+
+thread_local! {
+    static BUILDER: RefCell<FlatBufferBuilder<'static>> =
+        RefCell::new(FlatBufferBuilder::with_capacity(128));
+}
+
+pub trait FlatbufferSerializable<TInput>: Follow<'static> + Sized {
+    fn as_flat_buffer(builder: &mut FlatBufferBuilder<'static>, input: TInput) -> WIPOffset<Self>;
+
+    fn build_response(input: TInput) -> Vec<u8> {
+        BUILDER.with(|cell| {
+            let mut builder = cell.borrow_mut();
+            builder.reset();
+
+            let offset = Self::as_flat_buffer(&mut builder, input);
+
+            builder.finish(offset, None);
+            builder.finished_data().to_vec()
+        })
+    }
+}
+
+impl FlatbufferSerializable<Result<Option<bool>, &str>> for Response<'static> {
+    fn as_flat_buffer(
+        builder: &mut FlatBufferBuilder<'static>,
+        from: Result<Option<bool>, &str>,
+    ) -> WIPOffset<Response<'static>> {
+        let error_offset = from.as_ref().err().map(|e| builder.create_string(e));
+
+        let mut response_builder = ResponseBuilder::new(builder);
+
+        match from {
+            Ok(Some(flag)) => {
+                response_builder.add_enabled(flag);
+                response_builder.add_has_enabled(true);
+            }
+            Ok(None) => {
+                response_builder.add_has_enabled(false);
+            }
+            Err(_) => {
+                response_builder.add_error(error_offset.unwrap());
+            }
+        }
+
+        response_builder.finish()
+    }
+}
+
+impl FlatbufferSerializable<Result<Option<ExtendedVariantDef>, &str>> for Variant<'static> {
+    fn as_flat_buffer(
+        builder: &mut FlatBufferBuilder<'static>,
+        input: Result<Option<ExtendedVariantDef>, &str>,
+    ) -> WIPOffset<Self> {
+        let variant = input.unwrap();
+
+        if let Some(variant) = variant {
+            let payload_offset = variant.payload.as_ref().map(|payload| {
+                let payload_type_offset = builder.create_string(&payload.payload_type);
+                let value_offset = builder.create_string(&payload.value);
+
+                let mut variant_payload = VariantPayloadBuilder::new(builder);
+                variant_payload.add_payload_type(payload_type_offset);
+                variant_payload.add_value(value_offset);
+
+                variant_payload.finish()
+            });
+
+            let variant_name_offset = builder.create_string(&variant.name);
+
+            let mut variant_builder = VariantBuilder::new(builder);
+            variant_builder.add_feature_enabled(variant.feature_enabled);
+            variant_builder.add_enabled(variant.enabled);
+            variant_builder.add_name(variant_name_offset);
+            if let Some(payload_offset) = payload_offset {
+                variant_builder.add_payload(payload_offset);
+            }
+
+            variant_builder.finish()
+        } else {
+            let resp_builder = VariantBuilder::new(builder);
+            resp_builder.finish()
+        }
+    }
+}
+
+impl FlatbufferSerializable<Option<MetricBucket>> for MetricsBucket<'static> {
+    fn as_flat_buffer(
+        builder: &mut FlatBufferBuilder<'static>,
+        metrics: Option<MetricBucket>,
+    ) -> WIPOffset<Self> {
+        if let Some(metrics) = metrics {
+            let items: Vec<_> = metrics
+                .toggles
+                .iter()
+                .map(|(toggle_key, stats)| {
+                    let variant_items: Vec<_> = stats
+                        .variants
+                        .iter()
+                        .map(|(variant_key, count)| {
+                            let variant_key = builder.create_string(variant_key);
+                            let mut variant_builder = VariantEntryBuilder::new(builder);
+                            variant_builder.add_key(variant_key);
+                            variant_builder.add_value(*count);
+                            variant_builder.finish()
+                        })
+                        .collect();
+                    let variant_vector = builder.create_vector(&variant_items);
+
+                    let toggle_key = builder.create_string(toggle_key);
+                    let mut toggle_builder = ToggleStatsBuilder::new(builder);
+                    toggle_builder.add_no(stats.no);
+                    toggle_builder.add_yes(stats.yes);
+                    toggle_builder.add_variants(variant_vector);
+                    let toggle_value = toggle_builder.finish();
+                    let mut toggle_entry_builder = ToggleEntryBuilder::new(builder);
+                    toggle_entry_builder.add_value(toggle_value);
+                    toggle_entry_builder.add_key(toggle_key);
+                    toggle_entry_builder.finish()
+                })
+                .collect();
+            let toggle_vector = builder.create_vector(&items);
+            let mut resp_builder = MetricsBucketBuilder::new(builder);
+            resp_builder.add_start(metrics.start.timestamp_millis());
+            resp_builder.add_stop(metrics.stop.timestamp_millis());
+            resp_builder.add_toggles(toggle_vector);
+            resp_builder.finish()
+        } else {
+            let resp_builder = MetricsBucketBuilder::new(builder);
+            resp_builder.finish()
+        }
+    }
+}
+
+impl FlatbufferSerializable<Vec<ToggleDefinition>> for FeatureDefs<'static> {
+    fn as_flat_buffer(
+        builder: &mut FlatBufferBuilder<'static>,
+        known_toggles: Vec<ToggleDefinition>,
+    ) -> WIPOffset<Self> {
+        let items: Vec<_> = known_toggles
+            .iter()
+            .map(|toggle| {
+                let toggle_name_offset = builder.create_string(&toggle.name);
+                let project_offset = builder.create_string(&toggle.project);
+                let feature_type_offset = toggle
+                    .feature_type
+                    .as_ref()
+                    .map(|f| builder.create_string(f));
+
+                let mut feature_def_builder = FeatureDefBuilder::new(builder);
+
+                feature_def_builder.add_name(toggle_name_offset);
+                feature_def_builder.add_project(project_offset);
+                feature_def_builder.add_enabled(toggle.enabled);
+                if feature_type_offset.is_some() {
+                    feature_def_builder.add_type_(feature_type_offset.unwrap());
+                }
+                feature_def_builder.finish()
+            })
+            .collect();
+
+        let toggle_vector = builder.create_vector(&items);
+
+        let mut resp_builder = FeatureDefsBuilder::new(builder);
+        resp_builder.add_items(toggle_vector);
+        resp_builder.finish()
+    }
+}
